+++ samples/server/petstore/java-vertx/rx/src/main/java/io/swagger/server/api/verticle/StoreApiVerticle.java
@@ -40,7 +40,12 @@ public class StoreApiVerticle extends AbstractVerticle {
         //Consumer for deleteOrder
         vertx.eventBus().<JsonObject> consumer(DELETEORDER_SERVICE_ID).handler(message -> {
             try {
-                String orderId = message.body().getString("orderId");
+                String orderIdParam = message.body().getString("orderId");
+                if(orderIdParam == null) {
+                    manageError(message, new MainApiException(400, "orderId is required"), serviceId);
+                    return;
+                }
+                String orderId = orderIdParam;
                 service.deleteOrder(orderId).subscribe(
                     () -> {
                         message.reply(null);
@@ -73,7 +78,12 @@ public class StoreApiVerticle extends AbstractVerticle {
         //Consumer for getOrderById
         vertx.eventBus().<JsonObject> consumer(GETORDERBYID_SERVICE_ID).handler(message -> {
             try {
-                Long orderId = Json.mapper.readValue(message.body().getString("orderId"), Long.class);
+                String orderIdParam = message.body().getString("orderId");
+                if(orderIdParam == null) {
+                    manageError(message, new MainApiException(400, "orderId is required"), serviceId);
+                    return;
+                }
+                Long orderId = Json.mapper.readValue(orderIdParam, Long.class);
                 service.getOrderById(orderId).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
@@ -90,7 +100,12 @@ public class StoreApiVerticle extends AbstractVerticle {
         //Consumer for placeOrder
         vertx.eventBus().<JsonObject> consumer(PLACEORDER_SERVICE_ID).handler(message -> {
             try {
-                Order body = Json.mapper.readValue(message.body().getJsonObject("body").encode(), Order.class);
+                JsonObject bodyParam = message.body().getJsonObject("body");
+                if (bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                Order body = Json.mapper.readValue(bodyParam.encode(), Order.class);
                 service.placeOrder(body).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
