+++ samples/server/petstore/java-vertx/rx/src/main/java/io/swagger/server/api/verticle/UserApiVerticle.java
@@ -44,7 +44,12 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUser
         vertx.eventBus().<JsonObject> consumer(CREATEUSER_SERVICE_ID).handler(message -> {
             try {
-                User body = Json.mapper.readValue(message.body().getJsonObject("body").encode(), User.class);
+                JsonObject bodyParam = message.body().getJsonObject("body");
+                if (bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                User body = Json.mapper.readValue(bodyParam.encode(), User.class);
                 service.createUser(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -61,8 +66,13 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUsersWithArrayInput
         vertx.eventBus().<JsonObject> consumer(CREATEUSERSWITHARRAYINPUT_SERVICE_ID).handler(message -> {
             try {
-                List<User> body = Json.mapper.readValue(message.body().getJsonArray("body").encode(),
-                        Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
+                JsonArray bodyParam = message.body().getJsonArray("body");
+                if(bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                List<User> body = Json.mapper.readValue(bodyParam.encode(),
+                    Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
                 service.createUsersWithArrayInput(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -79,8 +89,13 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUsersWithListInput
         vertx.eventBus().<JsonObject> consumer(CREATEUSERSWITHLISTINPUT_SERVICE_ID).handler(message -> {
             try {
-                List<User> body = Json.mapper.readValue(message.body().getJsonArray("body").encode(),
-                        Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
+                JsonArray bodyParam = message.body().getJsonArray("body");
+                if(bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                List<User> body = Json.mapper.readValue(bodyParam.encode(),
+                    Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
                 service.createUsersWithListInput(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -97,7 +112,12 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for deleteUser
         vertx.eventBus().<JsonObject> consumer(DELETEUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
                 service.deleteUser(username).subscribe(
                     () -> {
                         message.reply(null);
@@ -114,7 +134,12 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for getUserByName
         vertx.eventBus().<JsonObject> consumer(GETUSERBYNAME_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
                 service.getUserByName(username).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
@@ -131,8 +156,18 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for loginUser
         vertx.eventBus().<JsonObject> consumer(LOGINUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
-                String password = message.body().getString("password");
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
+                String passwordParam = message.body().getString("password");
+                if(passwordParam == null) {
+                    manageError(message, new MainApiException(400, "password is required"), serviceId);
+                    return;
+                }
+                String password = passwordParam;
                 service.loginUser(username, password).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
@@ -165,8 +200,18 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for updateUser
         vertx.eventBus().<JsonObject> consumer(UPDATEUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
-                User body = Json.mapper.readValue(message.body().getJsonObject("body").encode(), User.class);
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
+                JsonObject bodyParam = message.body().getJsonObject("body");
+                if (bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                User body = Json.mapper.readValue(bodyParam.encode(), User.class);
                 service.updateUser(username, body).subscribe(
                     () -> {
                         message.reply(null);
